Incorporate JRuby implementation
This implementation is based on the JDK's
ScheduledThreadPoolExecutor class, which provides an efficient way
to schedule many tasks in the future without spinning up a thread
for each. This logic is ported from a Java implementation of
timeout that JRuby has shipped for many years, and passes all but
one of the timeout tests.