Terminate the timeout thread at exit
In some runtimes, the process may not go away but the Ruby runtime
should be shut down. This thread does not clean itself up, which
leads to such runtimes leaning on a best-effort kill+join of the
dangling threads late in the shutdown process. As a result these
threads may linger slightly after shutdown and fail expectations
that they will have completed and finalized.

This patch adds an `at_exit` for the timeout library to terminate
its own thread, which would happen earlier in the shutdown process
(`at_exit` execution is considered part of the "user" shutdown
phase in JRuby). This appears to fix issues running JRuby's
embedding API tests that check for timeout threads to have been
cleanly terminated and finalized.