Allow restarting timeout thread and queue
If the shutdown at_exit runs before other at_exit or termination
logic that still depends on the timeout thread, those subsequent
uses of timeout may fail. This additional patch allows the timeout
thread and queue to be restarted after the at_exit has fired, with
the side effect that it will not try to shutdown a second time.

This addresses review comments from #21.