Fix nested timeouts
ExitException#exception needs to always return self, so the same
exception object will be used.  The previous code never used the
same exception object, but the conditional was inverted from what
it should have been, so all existing tests passed.  There wasn't
a test for nested timeouts, so I added one.

Remove the @thread instance variable setting.  When using
Thread#raise, it is expected that the Thread that raises the
exception will be different than the thread that created the
exception.